Feature/increment count gpt requests

+++ bot/database/main.py
@@ -1,13 +1,17 @@
 from sqlalchemy.ext.asyncio import AsyncEngine, async_sessionmaker, create_async_engine
 
+from bot.misc.conf import settings
+
 
 def create_engine(db_uri: str) -> AsyncEngine:
-    return create_async_engine(db_uri, echo=True, future=True)
+    echo = True if settings.DEBUG else False
+    return create_async_engine(db_uri, echo=echo, future=True)
 
 
 async def proceed_schemas(engine: AsyncEngine, metadata) -> None:
     async with engine.begin() as conn:
-        await conn.run_sync(metadata.drop_all)
+        if settings.DEBUG:
+            await conn.run_sync(metadata.drop_all)
         await conn.run_sync(metadata.create_all)
